+++ utipy/pandas/resemble.py
@@ -2,35 +2,38 @@
 @author: ludvigolsen
 """
 
+from typing import Union
 from utipy.utils import _extended_describe
 import numpy as np
+import pandas as pd
 from random import shuffle
 
 
-def resemble(Series, distribution='uniform'):
+def resemble(
+    x: Union[pd.Series, np.ndarray, list],
+    distribution: str = 'uniform'
+) -> pd.Series:
     """
-    Generate resembling Series
+    Generate pandas.Series that resembles `x`.
 
     Generates data that resembles the original data by using descriptors
-    of the Series to create and sample from a specified distribution.
+    of `x` to create and sample from a specified distribution.
 
 
     Parameters
     ----------
-    Series : pd.Series
-        The Series to resemble.
+    x : pandas.Series, numpy.ndarray, list
+        The values to resemble.
     distribution : str
-        Distribution to sample from.
+        Distribution to sample from. One of:
             'uniform'
-                between min. and max.
+                Between min. and max.
             'gaussian'
-                from mean and std.
+                From mean and std.
             'robust gaussian'
-                from median and IQR.
-            'poisson'
-                NOT IMPLEMENTED PROPERLY YET.
+                From median and IQR.
             'shuffle'
-                shuffles original data.
+                Shuffles original data.
 
 
     Returns
@@ -46,24 +49,39 @@ def resemble(Series, distribution='uniform'):
 
     # TODO Check for NaNs
 
-    # Get description of Series
-    desc = _extended_describe(Series)
+    # Get description of `x`
+    desc = _extended_describe(x)
 
     if distribution == 'uniform':
         generated = np.random.uniform(
-            low=desc['min'], high=desc['max'], size=int(desc['count']))
+            low=desc['min'],
+            high=desc['max'],
+            size=int(desc['count'])
+        )
+
     elif distribution == 'gaussian':
         generated = np.random.normal(
-            loc=desc['mean'], scale=desc['std'], size=int(desc['count']))
+            loc=desc['mean'],
+            scale=desc['std'],
+            size=int(desc['count'])
+        )
+
     elif distribution == 'robust gaussian':
         generated = np.random.normal(
-            loc=desc['median'], scale=desc['IQR'], size=int(desc['count']))
+            loc=desc['median'],
+            scale=desc['IQR'],
+            size=int(desc['count'])
+        )
+
     elif distribution == 'poisson':
         # TODO Check up on this one. Should there be a min/max?
         generated = np.random.poisson(
-            lam=desc['max'], size=int(desc['count']))
+            lam=desc['max'],
+            size=int(desc['count'])
+        )
+
     elif distribution == 'shuffle':
-        generated = shuffle(Series)
+        generated = shuffle(x)
 
     # Change back to original dtype
     generated = generated.astype(desc['dtype'])