+++ utipy/pandas/subset_by_levels.py
@@ -4,12 +4,18 @@
 @author: ludvigolsen
 """
 
+from typing import List
 import pandas as pd
 
 # TODO Could this not be easily done with group_by -> split?
 # Consider refactoring or removing this function
 
-def subset_by_levels(data, cat_col, drop_cat_col=False):
+
+def subset_by_levels(
+    data: pd.DataFrame,
+    cat_col: str,
+    drop_cat_col: bool = False
+) -> List[pd.DataFrame]:
     """
     Subsets dataframe by each level of a categorical column.
 
@@ -32,12 +38,10 @@ def subset_by_levels(data, cat_col, drop_cat_col=False):
     Examples
     --------
 
-    Uncomment code to run.
-
-    # df = pd.DataFrame({'a': [1,2,3,4],
-    #                    'b': [2,3,4,5],
-    #                    'c': ['a','b','a','b']})
-    # subsets = subset_by_levels(df, cat_col = 'c')
+    >>> df = pd.DataFrame({'a': [1,2,3,4],
+    ...                    'b': [2,3,4,5],
+    ...                    'c': ['a','b','a','b']})
+    >>> subset_by_levels(df, cat_col = 'c')
 
     """