+++ utipy/utils/extended_describe.py
@@ -2,18 +2,20 @@
 @author: ludvigolsen
 """
 
+from typing import Union
+import pandas as pd
 import numpy as np
 from utipy.measures.iqr import _iqr
 
 # For pandas objects
 
-def _extended_describe(Series):
+def _extended_describe(x:Union[pd.Series, np.ndarray, list]) -> dict:
     """
     Adds median and IQR to `pandas.Series.describe()`.
     Only works for numeric series.
     """
-    desc = Series.describe()
-    desc['median'] = np.median(Series)
-    desc['IQR'] = _iqr(Series)
-    return desc
+    desc = pd.Series(x).describe()
+    desc['median'] = np.median(x)
+    desc['IQR'] = _iqr(x)
+    return dict(desc)